feat(discount): setup docker

+++ Services/Discount/Discount.Application/DTOs/CouponDto.cs
@@ -1,4 +1,4 @@
 namespace Discount.Application.DTOs
 {
-	public record CouponDto(int Id, string ProductName, string Description, decimal Amount);
+	public record CouponDto(int Id, string ProductName, string Description, int Amount);
 }
